+++ src/exceptions/file-already-exists.js
@@ -1,0 +1,8 @@
+export default class FileAlreadyExists extends Error {
+  constructor(file) {
+    super(`The "${file}" already exists
+    `);
+    this.name = 'FILEALREADYEXISTS';
+    this.code = 'FILEEX';
+  }
+}