+++ src/exceptions/index.js
@@ -4,7 +4,7 @@ import NoConsumerFoundException from './no-consumer-found';
 import MultipleScopesNoDefException from './multiple-scopes-no-def';
 import PackageJsonNotFound from './package-json-not-found';
 import PackageJsonAlreadyExists from './package-json-already-exists';
-
+import FileAlreadyExists from './file-already-exists';
 export {
   ComponentNotExistsException,
   ComponentNotExistsInException,
@@ -12,4 +12,5 @@ export {
   MultipleScopesNoDefException,
   PackageJsonNotFound,
   PackageJsonAlreadyExists,
+  FileAlreadyExists
 };