+++ src/index.js
@@ -1,6 +1,7 @@
 import 'regenerator-runtime/runtime';
 // import { bindAction, bindSpecificComponentsAction } from './actions';
 import { getDependencyTree } from './dependency-builder';
+import { pack } from './pack';
 import PackageJson from './package-json/package-json';
 
 // const lifecycleHooks = {
@@ -18,5 +19,6 @@ module.exports = {
   // bindSpecificComponents: bindSpecificComponentsAction,
   getDependencyTree,
   PackageJson,
+  pack,
 };