+++ src/pack/index.js
@@ -1,0 +1,5 @@
+import pack from './pack';
+
+export {
+  pack
+};